chore: refactor chat context; migrate to zustand for state management and remove obsolete context
