+++ src/features/chat/components/chat-input-box.tsx
@@ -3,10 +3,10 @@ import { useRef, useState } from "react"
 import SettingsButton from "@/components/settings-button"
 import { Textarea } from "@/components/ui/textarea"
 import { useAutoResizeTextarea } from "@/hooks/use-auto-resize-textarea"
-import { useLoadStream } from "@/context/load-stream-context"
 import { cn } from "@/lib/utils"
 import SendOrStopButton from "@/features/chat/components/send-or-stop-button"
-import { useChatInput } from "@/features/chat/context/chat-input-context"
+import { useChatInput } from "@/features/chat/hooks/use-chat-input"
+import { useLoadStream } from "@/features/chat/hooks/use-load-stream"
 import ModelMenu from "@/features/model/components/model-menu"
 import PromptSelectorDialog from "@/features/prompt/components/prompt-selector-dialog"
 import TabsSelect from "@/features/tabs/components/tabs-select"
