+++ src/features/chat/components/chat.tsx
@@ -1,8 +1,8 @@
 import { ScrollArea } from "@/components/ui/scroll-area"
-import { useLoadStream } from "@/context/load-stream-context"
 import ChatInputBox from "@/features/chat/components/chat-input-box"
 import ChatMessageBubble from "@/features/chat/components/chat-message-bubble"
 import { useChat } from "@/features/chat/hooks/use-chat"
+import { useLoadStream } from "@/features/chat/hooks/use-load-stream"
 import { OllamaStatusIndicator } from "@/features/model/components/ollama-status-indicator"
 import ChatSessionSelector from "@/features/sessions/components/chat-session-selector"
 import { useChatSessions } from "@/features/sessions/context/chat-session-context"
