+++ src/features/chat/hooks/use-chat-input.tsx
@@ -1,0 +1,7 @@
+import { chatInputStore } from "@/features/chat/stores/chat-input-store"
+
+export const useChatInput = () => {
+  const input = chatInputStore((s) => s.input)
+  const setInput = chatInputStore((s) => s.setInput)
+  return { input, setInput }
+}
