+++ src/features/chat/stores/chat-input-store.ts
@@ -1,0 +1,11 @@
+import { create } from "zustand"
+
+interface ChatInput {
+  input: string
+  setInput: (value: string) => void
+}
+
+export const chatInputStore = create<ChatInput>((set) => ({
+  input: "",
+  setInput: (text) => set({ input: text })
+}))
